+++ sql/core/src/test/resources/sql-tests/results/join.sql.out
@@ -1,5 +1,5 @@
 -- Automatically generated by SQLQueryTestSuite
--- Number of queries: 31
+-- Number of queries: 34
 
 
 -- !query 0
@@ -522,3 +522,155 @@ struct<key:int,key:int,key:int>
 2	2	2
 3	3	3
 4	4	4
+
+
+-- !query 31
+SELECT x.key, x.value, y.key, y.value, count(1) FROM
+  (SELECT * FROM testData UNION ALL
+   SELECT * FROM testData UNION ALL
+   SELECT * FROM testData UNION ALL
+   SELECT * FROM testData) x JOIN
+  (SELECT * FROM testData UNION ALL
+   SELECT * FROM testData UNION ALL
+   SELECT * FROM testData UNION ALL
+   SELECT * FROM testData) y
+WHERE x.key = y.key group by x.key, x.value, y.key, y.value
+-- !query 31 schema
+struct<key:int,value:string,key:int,value:string,count(1):bigint>
+-- !query 31 output
+1	1	1	1	16
+10	10	10	10	16
+100	100	100	100	16
+11	11	11	11	16
+12	12	12	12	16
+13	13	13	13	16
+14	14	14	14	16
+15	15	15	15	16
+16	16	16	16	16
+17	17	17	17	16
+18	18	18	18	16
+19	19	19	19	16
+2	2	2	2	16
+20	20	20	20	16
+21	21	21	21	16
+22	22	22	22	16
+23	23	23	23	16
+24	24	24	24	16
+25	25	25	25	16
+26	26	26	26	16
+27	27	27	27	16
+28	28	28	28	16
+29	29	29	29	16
+3	3	3	3	16
+30	30	30	30	16
+31	31	31	31	16
+32	32	32	32	16
+33	33	33	33	16
+34	34	34	34	16
+35	35	35	35	16
+36	36	36	36	16
+37	37	37	37	16
+38	38	38	38	16
+39	39	39	39	16
+4	4	4	4	16
+40	40	40	40	16
+41	41	41	41	16
+42	42	42	42	16
+43	43	43	43	16
+44	44	44	44	16
+45	45	45	45	16
+46	46	46	46	16
+47	47	47	47	16
+48	48	48	48	16
+49	49	49	49	16
+5	5	5	5	16
+50	50	50	50	16
+51	51	51	51	16
+52	52	52	52	16
+53	53	53	53	16
+54	54	54	54	16
+55	55	55	55	16
+56	56	56	56	16
+57	57	57	57	16
+58	58	58	58	16
+59	59	59	59	16
+6	6	6	6	16
+60	60	60	60	16
+61	61	61	61	16
+62	62	62	62	16
+63	63	63	63	16
+64	64	64	64	16
+65	65	65	65	16
+66	66	66	66	16
+67	67	67	67	16
+68	68	68	68	16
+69	69	69	69	16
+7	7	7	7	16
+70	70	70	70	16
+71	71	71	71	16
+72	72	72	72	16
+73	73	73	73	16
+74	74	74	74	16
+75	75	75	75	16
+76	76	76	76	16
+77	77	77	77	16
+78	78	78	78	16
+79	79	79	79	16
+8	8	8	8	16
+80	80	80	80	16
+81	81	81	81	16
+82	82	82	82	16
+83	83	83	83	16
+84	84	84	84	16
+85	85	85	85	16
+86	86	86	86	16
+87	87	87	87	16
+88	88	88	88	16
+89	89	89	89	16
+9	9	9	9	16
+90	90	90	90	16
+91	91	91	91	16
+92	92	92	92	16
+93	93	93	93	16
+94	94	94	94	16
+95	95	95	95	16
+96	96	96	96	16
+97	97	97	97	16
+98	98	98	98	16
+99	99	99	99	16
+
+
+-- !query 32
+SeleCT * from
+  (select * from upperCaseData WherE N <= 4) leftTable fuLL OUtER joiN
+  (sElEcT * FROM upperCaseData whERe N >= 3) rightTable
+    oN leftTable.N = rightTable.N
+-- !query 32 schema
+struct<N:int,L:string,N:int,L:string>
+-- !query 32 output
+1	A	NULL	NULL
+2	B	NULL	NULL
+3	C	3	C
+4	D	4	D
+NULL	NULL	5	E
+NULL	NULL	6	F
+
+
+-- !query 33
+SELECT * FROM src1 as a JOIN src1 as b on a.value <=> b.value
+-- !query 33 schema
+struct<key:int,value:string,key:int,value:string>
+-- !query 33 output
+165	NULL	165	NULL
+165	NULL	201	NULL
+165	NULL	NULL	NULL
+165	val_165	165	val_165
+201	NULL	165	NULL
+201	NULL	201	NULL
+201	NULL	NULL	NULL
+330	val_330	330	val_330
+86	val_86	86	val_86
+NULL	NULL	165	NULL
+NULL	NULL	201	NULL
+NULL	NULL	NULL	NULL
+NULL	val_null	NULL	val_null
